refactor: 構文解析ライブラリを lib/parser/ に集約
parser / ambiguous_expr / cfg_graph の3ヘッダを lib/string/ から独立した
lib/parser/ へ移動。docs も docs/parser/ に移し documentation_of と include
例を追随、逆依存の9テスト (AOJ 0109/0264/1102/1244/2255/2570/2584/2613/2704)
の include も更新。全テストの syntax-check 済み。
