+++ test/aoj/challenge/0264.test.cpp
@@ -3,7 +3,7 @@
 #include <iostream>
 #include <stdexcept>
 #include <string>
-#include "string/parser.hpp"
+#include "parser/parser.hpp"
 
 // 有限体電卓: 素数 p の有限体 GF(p) 上で四則演算式を評価する。
 // 除数が p の倍数 (≡0) のときは NG。逆元はフェルマーの小定理で y^(p-2)。
